requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas
streamlit
plotly
scikit-learn
numpy
folium
streamlit-folium
